audio_transcriber/requirements.txt
deepgram-sdk
pyaudio
websocket-client
httpx
supabase
sounddevice 
pydub 
scipy 
numpy